+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
 
+## [v5.5.0] - 2020-02-28
+
 ### Fixed
 
 - `api`
@@ -269,7 +271,9 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 - Improve README.md documentation.
 - Changed the getUserAuth and getOAuth2Token to use the new API auth functions.
 
-[unreleased]: https://github.com/itslanguage/itslanguage-js/compare/v5.4.0...HEAD
+[unreleased]: https://github.com/itslanguage/itslanguage-js/compare/v5.5.0...HEAD
+[v5.5.0]: https://github.com/itslanguage/itslanguage-js/compare/v5.4.1...v5.5.0
+[v5.4.1]: https://github.com/itslanguage/itslanguage-js/compare/v5.4.0...v5.4.1
 [v5.4.0]: https://github.com/itslanguage/itslanguage-js/compare/v5.3.0...v5.4.0
 [v5.3.0]: https://github.com/itslanguage/itslanguage-js/compare/v5.2.0...v5.3.0
 [v5.2.0]: https://github.com/itslanguage/itslanguage-js/compare/v5.1.0...v5.2.0